moved _array_api_for_tests
